+++ packages/ui-inputs/src/index.ts
@@ -1,4 +1,4 @@
-export const VERSION = '0.1.0' as const;
+export const VERSION = '0.1.1' as const;
 
 export {
   VarshylTextInput,